refactor: CRUD 기능 DB Helper로 이동
